feat: add movement capability
Add #move to Board to allow movement of pieces on the board
Add #empty to Board to check if given cell is empty

Add #generate_moves to all children of Piece to return an array of the different
ways they can move.
Add #valid_move? to determine if a given move is still within bounds, and if it's
an acceptable move for the particular piece being moved

Add #display_column_letters to make it easier to identify cells
Modify #display to display row numbers for the same reason

Add tests for #generate_moves functions, and #valid_move?